Move Django test app into a testapp directory to avoid import problems importing 'tests' before starting up Django.
