+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
-addopts = --cov=braces
+addopts = --cov=braces --ds=tests.settings
 django_find_project = false
