+++ tests/test_forms.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 from django import test
 from django.contrib.auth.models import User
 
-from . import forms
+from tests.testapp import forms
 
 
 class TestUserKwargModelFormMixin(test.TestCase):
