+++ tox.ini
@@ -9,8 +9,7 @@ basepython =
     py37: python3.7
     py38: python3.8
 
-commands =
-	{posargs:pytest}
+commands = python -m pytest {posargs}
 
 deps =
     mock
